requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pymongo==4.5.0
flask_pymongo==2.3.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.0
google-ai-generativelanguage==0.4.0
pdfplumber==0.10.0
pdfminer.six==20221105
python-docx==0.8.11
PyJWT==2.8.0
scikit-learn==1.3.0
numpy==1.23.5
nltk==3.8.1
PyPDF2==3.0.1
pydantic==2.7.1
spacy==3.8.4

